requirements.txt
# Nova AI Backend Requirements

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# LLM Orchestration
dspy-ai>=2.4.0
litellm>=1.44.0

# Vector Database
qdrant-client>=1.6.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# Voice (if using STT/TTS)
# Uncomment if you need these:
# speechrecognition>=3.10.0
# pyttsx3>=2.90
# pyaudio>=0.2.13

# Optional: For debugging LiteLLM
# Uncomment to enable debug logging:
# openai>=1.3.0
